combine both empty states and existed card page under one wrapper
